+++ src/Futhark/CLI/Eval.hs
@@ -7,9 +7,10 @@ import Control.Monad
 import Data.Map qualified as M
 import Data.Text qualified as T
 import Futhark.Eval
-  ( InterpreterConfig (..),
-    interpreterConfig,
-    newFutharkiState,
+  ( EvalConfig (..),
+    evalConfig,
+    evalServerOptions,
+    newInterpreterState,
     runExpr,
   )
 import Futhark.Util.Options
@@ -27,15 +28,14 @@ import System.IO (stderr)
 
 -- | Run @futhark eval@.
 main :: String -> [String] -> IO ()
-main = mainWithOptions interpreterConfig options "options... <exprs...>" run
+main = mainWithOptions evalConfig options "options... <exprs...>" run
   where
     run [] _ = Nothing
     run exprs config = Just $ runExprs exprs config
 
-runExprs :: [String] -> InterpreterConfig -> IO ()
+runExprs :: [String] -> EvalConfig -> IO ()
 runExprs exprs cfg = do
-  let InterpreterConfig _ file = cfg
-  maybe_new_state <- newFutharkiState cfg file M.empty
+  maybe_new_state <- newInterpreterState cfg M.empty
   interpreter_state <- case maybe_new_state of
     Left reason -> do
       hPutDocLn stderr reason
@@ -43,21 +43,20 @@ runExprs exprs cfg = do
     Right s -> pure s
   forM_ exprs $ \expr -> putDocLn =<< runExpr interpreter_state (T.pack expr)
 
-options :: [FunOptDescr InterpreterConfig]
+options :: [FunOptDescr EvalConfig]
 options =
   [ Option
       "f"
       ["file"]
       ( ReqArg
-          ( \entry -> Right $ \config ->
-              config {interpreterFile = Just entry}
-          )
+          (\entry -> Right $ \config -> config {evalFile = Just entry})
           "NAME"
       )
       "The file to load before evaluating expressions.",
     Option
       "w"
       ["no-warnings"]
-      (NoArg $ Right $ \config -> config {interpreterPrintWarnings = False})
+      (NoArg $ Right $ \config -> config {evalPrintWarnings = False})
       "Do not print warnings."
   ]
+    ++ evalServerOptions